about negative bite: it kinda works; likes to fall on its face, though

+++ src/bite.cpp
@@ -17,7 +17,7 @@ Rcpp::List CPP_bite(const Rcpp::List& x,
     } else {
         throw std::invalid_argument("indices must be either all positive or all negative");
     }
-    auto ret = bite<RCPP_IT>(import_from_R(x, NA_letter), Rcpp::as<std::vector<int>>(cpp_indices));
+    auto ret = bite<RCPP_IT>(import_from_R(x, NA_letter), Rcpp::as<std::vector<long long int>>(cpp_indices));
     return Rcpp::List::create(Rcpp::Named("warning", std::get<0>(ret)),
                               Rcpp::Named("sq", export_to_R(std::get<1>(ret))));
 }
